requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
ultralytics==8.0.196
pillow==10.1.0
python-multipart==0.0.6</content>
<parameter name="filePath">d:\git\water_logging\requirements.txt